refactor(threads): consolidate sub-thread logic into threads module
Move sub-thread management from separate sub_threads.ts file into the
existing threads module structure. Business logic is now in
model/threads/get_or_create_sub_thread.ts with a thin wrapper in
threads.ts, following the established pattern.
