+++ services/platform/convex/agent_tools/sub_agents/helpers/get_or_create_sub_thread.ts
@@ -51,7 +51,7 @@ export async function getOrCreateSubThread(
 
   // Delegate to atomic internal mutation to prevent race conditions
   // The mutation handles read-check-create-update in a single transaction
-  return await ctx.runMutation(internal.sub_threads.getOrCreateSubThreadAtomic, {
+  return await ctx.runMutation(internal.threads.getOrCreateSubThreadAtomic, {
     parentThreadId,
     subAgentType,
     userId,
