+++ services/platform/convex/model/threads/index.ts
@@ -13,8 +13,13 @@ export { updateChatThread } from './update_chat_thread';
 export { getLatestThreadWithMessageCount } from './get_latest_thread_with_message_count';
 export { getLatestToolMessage } from './get_latest_tool_message';
 export { getThreadMessagesStreaming } from './get_thread_messages_streaming';
+export { getOrCreateSubThread } from './get_or_create_sub_thread';
 
 export type { ThreadMessage } from './get_thread_messages';
 export type { Thread } from './list_threads';
 export type { LatestToolMessage } from './get_latest_tool_message';
 export type { StreamingMessagesResult } from './get_thread_messages_streaming';
+export type {
+  SubAgentType,
+  GetOrCreateSubThreadResult,
+} from './get_or_create_sub_thread';
